+++ src/Caller/Service/CallerService.php
@@ -52,7 +52,7 @@ class CallerService implements CallerServiceInterface
             !isset($page->getInfo()['http_code']) ||
             200 !== (int) $page->getInfo()['http_code']
         ) {
-            throw Caller::byHttpResponseCode();
+            throw Caller::byHttpResponseCode((int) $page->getInfo()['http_code']);
         }
         if (0 < strlen($page->getErrorMessage())) {
             throw Caller::byErrorMessage($page->getErrorMessage());