+++ src/Report/Infrastructure/ReportProvider.php
@@ -18,6 +18,28 @@ class ReportProvider implements ReportProviderInterface
     {
     }
 
+    public function getLatestValidReport(): Report
+    {
+        $report = oxNew(Report::class);
+
+        $queryBuilder = $this->queryBuilderFactory->create();
+        $queryBuilder->select($report->getViewName() . '.oxid')
+            ->from($report->getViewName())
+            ->where('json_length(report) > 0')
+            ->orderBy($report->getViewName() . '.issued_at', 'DESC')
+            ->setMaxResults(1);
+
+        /** @var \Doctrine\DBAL\Statement $result */
+        $result = $queryBuilder->execute();
+        $resultId = (string) $result->fetchOne();
+
+        if ($resultId) {
+            $report->load($resultId);
+        }
+
+        return $report;
+    }
+
     public function getLatestReport(): Report
     {
         $report = oxNew(Report::class);