+++ tests/Codeception/Acceptance/FrontendCest.php
@@ -7,7 +7,7 @@
 
 declare(strict_types=1);
 
-namespace EyeAble\EyeAbleAssist\Tests\Tests\Codeception;
+namespace EyeAble\EyeAbleAssist\Tests\Codeception;
 
 use Codeception\Util\Fixtures;
 use EyeAble\EyeAbleAssist\Tests\Codeception\AcceptanceTester;