+++ tests/Codeception/AcceptanceAdmin/AdminCest.php
@@ -26,7 +26,6 @@ final class AdminCest
         $I->loginAdmin();
 
         $I->waitForText(Translator::translate('EYEABLE_REPORT_TITLE'));
-        $I->waitForText(Translator::translate('EYEABLE_REPORT_EXPLANATION'));
         $I->waitForText(Translator::translate('EYEABLE_PLEASE_WAIT_FOR_REPORT'));
     }
 
@@ -40,13 +39,10 @@ final class AdminCest
         $I->loginAdmin();
 
         $I->waitForText(Translator::translate('EYEABLE_REPORT_TITLE'));
-        $I->waitForText(Translator::translate('EYEABLE_REPORT_EXPLANATION'));
         $I->waitForText(Translator::translate('EYEABLE_LATEST_REPORT_PAGE') . 'http://myoxidehop.local');
         $I->waitForText(Translator::translate('EYEABLE_LATEST_REPORT_ERRORCOUNT') . '66');
-        $I->waitForText(
-            Translator::translate('EYEABLE_LATEST_REPORT_DATE') .
-            date('Y-m-d 12:12:12')
-        );
+        $I->waitForText(Translator::translate('EYEABLE_LATEST_REPORT_DATE'));
+        $I->waitForText(date('Y-m-d 12:12:12'));
     }
 
     private function insertReport(AcceptanceAdminTester $I): void