+++ tests/PhpStan/phpstan-bootstrap.php
@@ -8,16 +8,11 @@
 declare(strict_types=1);
 
 class_alias(
-    \OxidEsales\Eshop\Application\Model\User::class,
-    \EyeAble\EyeAbleAssist\Model\User_parent::class
-);
-
-class_alias(
     \OxidEsales\Eshop\Application\Controller\StartController::class,
     \EyeAble\EyeAbleAssist\Controller\StartController_parent::class
 );
 
 class_alias(
-    \OxidEsales\Eshop\Application\Model\Basket::class,
-    \EyeAble\EyeAbleAssist\Model\Basket_parent::class
+    \OxidEsales\Eshop\Core\ViewConfig::class,
+    \EyeAble\EyeAbleAssist\Shop\ViewConfig_parent::class
 );