+++ views/admin_twig/de/module_options.php
@@ -12,7 +12,7 @@ $aLang = [
     'EYEABLE_REPORT_TITLE' => 'Eye-Able® Kurzreport',
     'EYEABLE_REPORT_EXPLANATION' => '<b><a href="eye-able.com" style="color:#0809ff">Eye-Able® </a></b> prüft automatisiert Ihre Startseite alle 7 Tage auf Barrierefreiheitsfehler nach der WCAG 2.2. <br>Sie wollen mehr darüber wissen, einen detaillierten Bericht oder mehr Seiten scannen lassen? <br>Dann kontaktieren Sie uns gerne unter <b><a href="mailto:info@eye-able.com" style="color:#0809ff">info@eye-able.com</a></b>',
     'EYEABLE_LATEST_REPORT_PAGE' => 'Geprüfte Seite: ',
-    'EYEABLE_LATEST_REPORT_ERRORCOUNT' => 'Anzahl Fehler: ',
+    'EYEABLE_LATEST_REPORT_ERRORCOUNT' => 'Anzahl WCAG-Probleme: ',
     'EYEABLE_LATEST_REPORT_DATE' => 'Letzte Prüfung am: ',
     'EYEABLE_PLEASE_WAIT_FOR_REPORT' => 'Bitte warten, Ihre Eye-Able® Prüfung wird durchgeführt.',
     'SHOP_MODULE_GROUP_eyeableassist_main' => 'Einstellungen',