+++ views/admin_twig/en/module_options.php
@@ -12,7 +12,7 @@ $aLang = [
     'EYEABLE_REPORT_TITLE' => 'Eye-Able® teaser report',
     'EYEABLE_REPORT_EXPLANATION' => '<b><a href="eye-able.com/en" style="color:#0809ff">Eye-Able® </a></b> automatically checks your homepage every 7 days for accessibility errors according to WCAG 2.2. Would you like to know more about this, receive a detailed report or have more pages scanned? Then please contact us at <b><a href="mailto:info@eye-able.com" style="color:#0809ff">info@eye-able.com</a></b>',
     'EYEABLE_LATEST_REPORT_PAGE' => 'Page checked: ',
-    'EYEABLE_LATEST_REPORT_ERRORCOUNT' => 'Error count: ',
+    'EYEABLE_LATEST_REPORT_ERRORCOUNT' => 'WCAG issues count: ',
     'EYEABLE_LATEST_REPORT_DATE' => 'Last examination on: ',
     'EYEABLE_PLEASE_WAIT_FOR_REPORT' => 'Please wait some minutes while Eye-Able® Report is being created.',
     'SHOP_MODULE_GROUP_eyeableassist_main' => 'Settings',